Week7/python/practice/guardrails_practice/requirements.txt
langchain
langchain-groq
pydantic
streamlit
python-dotenv
guardrails-ai
